requirements.txt
# Elite Trading System - Python Dependencies (TESTED for Python 3.13)

# ============================================================================
# CORE
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
schedule==1.2.0

# ============================================================================
# DATA COLLECTION
# ============================================================================
yfinance==0.2.50
finvizfinance==1.2.0
pandas>=2.2.0
numpy>=1.26.0
requests==2.32.3

# ============================================================================
# WEB SCRAPING
# ============================================================================
selenium==4.27.1
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
lxml==5.3.0

# ============================================================================
# API / WEB FRAMEWORK
# ============================================================================
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-multipart==0.0.20

# ============================================================================
# FRONTEND
# ============================================================================
streamlit==1.41.1
plotly==5.24.1

# ============================================================================
# DATABASE
# ============================================================================
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# ============================================================================
# MACHINE LEARNING
# ============================================================================
scikit-learn==1.6.0
xgboost==2.1.3
joblib==1.4.2

# ============================================================================
# TIMEZONE
# ============================================================================
pytz==2024.2

# ============================================================================
# ALERTS
# ============================================================================
python-telegram-bot==21.9

# ============================================================================
# UTILITIES
# ============================================================================
tqdm==4.67.1
colorama==0.4.6
